Merge r1600 into stable branch.
------------------------------------------------------------------------
r1600 | keimpema | 2013-06-24 17:23:40 +0200 (Mon, 24 Jun 2013) | 6 lines

Round the number of ffts per integration down. In the old situation
the last fft of an integration could overlap with the next integration.
This means that in the final integation of a scan datapoints from after
the scan end time are needed. This triggers an assertion in the delay
model.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/branches/stable-3.0@1601 5736e3ea-d426-0410-8aaf-93e336ac3372