------------------------------------------------------------------------ r1893 | keimpema | 2015-05-21 14:59:55 +0200 (Thu, 21 May 2015) | 10 lines
Add option "exit_on_empty_datastream" to determine if the correlation should
be aborted when one of the input files contains no data. 

The default, as it has been for many years, is to abort the correlation to make sure
that the operators will notice ASAP when there is a faulty diskpack.

But for the automatic fringe tests this behaviour is unwanted. If one station
uploads a defective file you don't want to abort the fringe test and require human
intervention.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/branches/stable-3.3@1895 5736e3ea-d426-0410-8aaf-93e336ac3372