Add fixes to the averaging procedure to the new stable branch, also add support for the latest IPP ------------------------------------------------------------------------ r2042 | keimpema | 2017-11-28 16:50:59 +0100 (Tue, 28 Nov 2017) | 3 lines
Some of the buffers set by ippsFFTGetSize can also be zero,
take this into account.

------------------------------------------------------------------------
r2041 | keimpema | 2017-11-28 11:48:20 +0100 (Tue, 28 Nov 2017) | 2 lines

Oops, used ippsFree instead of ippFree in sfxc_fft

------------------------------------------------------------------------
r2040 | keimpema | 2017-11-28 10:09:25 +0100 (Tue, 28 Nov 2017) | 3 lines

Fix the FFT code so that it works with the latest version of IPP (IPP2018.1).
The new code still works with IPP7.1 and 8.1

------------------------------------------------------------------------
r2039 | kettenis | 2017-11-20 14:59:02 +0100 (Mon, 20 Nov 2017) | 2 lines

Remove some commented-out code.

------------------------------------------------------------------------
r2038 | keimpema | 2017-11-16 13:33:43 +0100 (Thu, 16 Nov 2017) | 8 lines

Fix the averaging down in frequency. The rectangular window is now
treated as special case. The is necessary because the periodicity
of this window function differs by one point as compared to the
Hann, Hamming, and Cosine window functions.

The second thing that was fixed is that the last point of the lag
spectrum was not normalized correctly (for all window functions).


git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/branches/stable-3.5@2047 5736e3ea-d426-0410-8aaf-93e336ac3372