Merge sample_rate patch from trunk ------------------------------------------------------------------------ r2188 | keimpema | 2019-07-08 15:44:17 +0200 (Mon, 08 Jul 2019) | 13 lines
In vex2 there will no longer be a sample_rate field in the $FREQ section,
instead this information is obtained from $TRACKS / $BITSTREAMS / $DATASTREAMS.

This requires the correlator to know what data format a station is using which
the correlator obtains from the $DAS section. However, when using a virtual
setup station (e.g. for mixed bandwith or zoom bands) we don't add it to
$STATION to not complicate the j2ms2 step, which there is no $DAS for
the reference station. This crashes the correlator.

As a temporary work around, for vex 1.5 files we return the sample_rate
from the $FREQ section without looking at $DAS.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/branches/stable-4.0@2189 5736e3ea-d426-0410-8aaf-93e336ac3372