Also apply input node fix to stable branch ------------------------------------------------------------------------ r2301 | keimpema | 2021-04-12 18:41:57 +0200 (Mon, 12 Apr 2021) | 12 lines
In Input_node_data_writer the list of integer delay changes
accidentally got copied with every write, which seriously
impacted input node performance.

Creating a copy of the integer delay list is necessary because we
add the per-channel delays in the data writer. As each writer handles
only channel, the per channel delay can simply be applied to the
integer delay list of each data writer.

The copy is now moved to fetch_next_time_interval, so that it is only
done once at the start of a new scan.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/branches/stable-5.1@2302 5736e3ea-d426-0410-8aaf-93e336ac3372